requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
Cerberus==1.1
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.2
django-celery==3.2.2
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
django-jet==1.0.7
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
kombu==3.0.37
Markdown==2.6.11
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
serpy==0.3.1
six==1.11.0
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1

